Fixed distance
Now it is properly turned on and off together with glassCore. This results in more ties between the original glass and the new features, which makes it harder to update glass in case IndeedPete pushes out an update. This sucks, but even with a completely modular approach some crossing over would have been inevitable and updating nearly as hard as it is now.